Overall update, data set included, gallery mode
This adds in the mixing of kun and on readings marked with an * (on) and also fixes the writing in to remove non typed characters. It also adds a significant change to the usage in that now hitting control will switch between gallery mode and quiz mode where in gallery mode all tiles display their meanings and pronunciations while quiz is how it was where you only get the symbol and have to quiz that info.

Still haven't fixed the fact that click one isn't shown right and also no example sentences yet but I'm burning to add them this weekend. Have school kind of all week so the next few days probably not but yeah. Some good improvements in here.

Oh yeah, I moved the text data set to the included files so now it's part of the transfered files. Good stuff.