Boom boom boom, example sentences done
All words now have an array of sentence indexs given by a big cross reference at the beginning. (this all takes about 15 seconds on my laptop, sorry, but I did make the core loading logic way faster with a simple change..) And when you're clicking through the kanji the example tag now contains displays of 1 of 13000 possible sentences, but only if that kanji is contained in the sentence. So instead of nothing you now get a click through of basically infinite sentences. Pretty cool. And I do actually have reading data in there already but I'm not sure how / if I have the energy to make a custom text displayer that shows the tagged elements above their source text. That's kinda the goal though, to have that data visible and all. Anyway, very interesting...