Fixed examples and switched a bunch of loading systems, faster
So the sentence tagging no longer happens up front and thus there's not nearly as much load in time, instead the words get individually checked when you ask for them in game which is 2500 times faster per request but happens every time. Also I switched to ds_lists because apparently they're faster for loading in.