Polish and bug fixes
The wheels are now bigger and the menus now have proper scaling and contain their text a little better, everything has been made smoother and bigger to fill the screen more completely. This also includes properish scaling variables so changing the size of the wheels and info menu should be as simple as changing two numbers but I didn't verify completely.
I also fixed the issue of the screen clipping on the edge and the fact that the first click would place you differently than the visual. I'm not honestly sure why that happens still, something about the size of the surface being rounded into tile sizes leading to target offsets... Anyways, it's fixed for now, if you change the scale of the camera or tiles it would probably be broken again but idk. But indeed, everything's a little nicer.
I still want to make map/environment generation/visuals and also add rubi text to the kanji to help pronounce unknown kanji and finally filters and map editing but the filters require word data that I don't have so...